(4.21) delete blog requires valid token

+++ part4/bloglist/controllers/blog.js
@@ -17,7 +17,7 @@ blogRouter.post('/', async (request, response) => {
     const body = request.body;
     const decodedToken = jwt.verify(request.token, process.env.SECRET);
     if (!request.token || !decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid'});
+        return response.status(401).json({ error: 'token missing or invalid' });
     }
     const user = await User.findById(decodedToken.id);
 
@@ -37,8 +37,20 @@ blogRouter.post('/', async (request, response) => {
 });
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' });
+    }
+    
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === user._id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+        return response.status(204).end();
+    }
+
+    return response.status(401).json({ error: 'unauthorized access'});
 });
 
 blogRouter.put('/:id', async (request, response) => {
